Replace `spawn` method with `exec` in exec_command function (#7)
Description
===========

The function `pub fn exec_command<C, I, S>(command: C, args: I) -> io::Result<i32>`
is used to call a child process (in most cases a sub command) and run
it. As of now this meant that the mother process `uvm` was still
executing. I replaces `spawn` with the system depending function `exec`.
It is save to call `exec` here because both usecases of `exec_command`
just gather resources to hand them over to the child process.

Changes
=======

![IMPROVE] sub command invocation by using `exec` instead of `spawn`