Add basic unity version manifest logic
Description
===========

This patch adds a first draft version to handle unity `ini` manifest
files. For each unity version released, unity publishes a platform
depended `ini` file to their S3 storage. This file describes all
components for a release. This file is intended to be used by the
platform installers but Unity Hub is using them too when installing from
the version archive. I'm not very happy with some implementation details
but want to close the circle first before attempting a second throw at
this issue. The most important function, get the installer url for a
given component, is already implemented and is also the main reason to
push this pull request and continue.

Changes
=======

![ADD] basic unity version manifest cache/loader