Fix windows installer logic (#15)
Description
===========

The windows unity installer accepts a commandline flag `/D` to set the
destination for the installation. This flag must be provided in the form
`/D=some\destination`. The rust `std::process::Command::spawn` method
does some things to the provides arguments so that the installer skips
over the flag. The result is that the installer uses the last selected
destination and writes over the possible installation.
This patch fixes this problem with a temp solution. I couldn't find a
incantation from rust that worked so the patch logic introduces a
tempary batch file with the correct installer incatation.

Changes
=======

![FIX] windows unity installer to destination