Add custom install downloader (#18)
* Add custom install downloader

Description
===========

The first implementation used `brew cask` to download the unity
module installers. This implementation doesn't work on windows or linux
so this patch adds a first implementation for a custom download logic.
The installer locations are fetched from the version manifest and
downloaded to a `cache` directory:

* macos: `~/Library/Caches/Wooga/UnityVersionManager/installers`
* windows : `%HOME%/AppData/Local/UnityVersionManager/installers`

The download process is restartable if the tool fails or any other
interrupt happens. The install command and the download function will
create a lock file to lock the temporary installation file so that only
one uvm process can access it at the same time.

All installer binaries are verified with the provided `md5` checksum
from the unity version manifest.

This implementation acts as a base and needs heavy refactoring.

Changes
=======

![ADD] unity installer download implementation
![ADD] process locking
![ADD] checksum verification for downloaded installer binaries

* Remove nightly linux checks

* Create utils module with custom process lock function