Remove brew support Part II: uninstall without brew
Description
===========

This is part II of a series of patches to remove brew dependency. The
`uninstall` command collected all installed `brew` casks matching a
specific naming pattern (`unity.*@{version}`) and matched them against
the provided options. One fault of this implementation is the
disconnection between `brew` cask state and the actual installation.
The new implementation will use the `uvm` installation and installation
components information to delete the requested editor or editor
components.

Changes
=======

![REMOVE] `brew` dependency in uninstall command