Remove brew support Part III: versions command without brew (#22)
Description
===========

This is part III of a series of patches to remove brew dependency. The
versions command used `brew` search command to fetch a list of available
casks with the pattern `unity@`. This patch will use the still static
versions list instead. This makes the command faster and smaller. The
internal versions list will be adjusted to fetch an up to date list of
available versions in a later patch.

Changes
=======

![REMOVE] brew dependency in versions command