Fix component install path and install location check (#26)
Description
===========

Windows installers only need the path to the directory which contains
the `Unity.exe`. But to check if a specific component is installed we
need also the complete path to the installation location.
This patch adds a another function to `Component` to fetch either the
installer path or the installation location. On mac os both paths are
equal.

Changes
=======

![ADD] new function to `Component` `install_location`
![IMPROVE] installer process by separating between installer path and install location