Use unity-versions-service for version lookup (#33)
Description
===========

Instead of using a static list of released versions, we should use a
remote service to fetch a recent list of released unity versions. This
patch adds a remote call to the [unity-versions-service] to fetch a
complete list of all available versions to install and to fetch the
revision hash for a given version to create the manifest url.

all versions
-------------

The function `uvm_core::unity::version::hash::all_versions` used to
create a iterator from a `HashMap` created at runtime from a static yml
string. This `HashMap` is still in place for now for the revision hash
fetch but no longer used for the list of all available versions.
The new implementation will call a remove service
[unity-versions-service] with the endpoint `/versions` to retrieve an
up-to-date list of all versions. There is no local cache or service
config in this patch. Every time the
`uvm_core::unity::version::hash::all_versions` function is called, the
remote server will be called as well. Future patches will change this.

version hashes
--------------

The function `uvm_core::unity::version::hash::hash_for_version` used to
fetch the revision hash for the given version from a semi static
`HashMap` (see _all versions_). This patch adds two fallback functions
that gets triggered if the given version is not included in `HashMap`.

1. the hash will be fetched from a local cache file
2. the [unity-versions-service] will be called at the endpoint `/versions/${VERSION}/hash`

If the hash was fetched from the remote service it will be saved in a
local cache file to limit network calls in the future.

Changes
=======

* ![IMPROVE] use [unity-versions-service] to fetch list of available versions
* ![IMPROVE] use [unity-versions-service] as a fallback to retrieve revision hash for a given versions

[unity-versions-service]: https://github.com/Larusso/unity-versions-service