Add error chains for better error handling (#35)
Description
===========

Error handling was not the major focus when starting this project. It is
now more important so this patch adds the crate [error_chains] to help
create more helpful errors and reasons attached.

At the moment most errors are converted from the inner error to a more
generic error

```rust
impl From<std::io::Error> for UvmError
```

This makes working with multiple error types in one method bearable but
isn't helpful for the logic calling these API's. The reason why and
what failed gets lost because the inner error gets bubbled up. And a
`entity not found` `std::io::Error` when installing a Unity version and
multiple components is not helpful. This is just the initial
implementation. There is room for improvement but I need to play around
with the crate first.

Changes
=======

* ![IMPROVE] error handling by introducing [error_chain]

[error_chain]:		https://crates.io/crates/error-chain