Fix windows installation logic (#36)
Description
===========

Fix install logic
-----------------

The windows install logic was still broken due to some path errors
introduced in a earlier commit. It was possible to install versions, but
the installer didn't use the passed destination value and either
installed to the default destination or to the last destination used.
Both editor and component installer behave that way.

Remove preinstalled components from component list
--------------------------------------------------

On windows the __windows__ component is preinstalled. That is the same
on mac os where the macos component is installed with the editor. The
function `installed_components` will no longer report the __windows__ or
__windows-mono__ component on windows.

Fix install location for components
-----------------------------------

The install location `Editor` is not working. The installer claims not
to find the __Unity.exe__ in the specified directory. Which is strange
since it resides in the `Editor` directory of the installation. The
component installer wants to have the root installation directory which
only contains the `Editor` directory (it seems that unity itself is not
100% sure what the programm root actually is)

Changes
=======

* ![FIX] windows install logic
* ![FIX] windows component list
* ![FIX] windows install location for components