Improve manifest loading and caching
Description
===========

The logic to download and cache unity version `.ini` files had some
issues. The first is that it saved and cached a file even though the
http response was not successful. The second issue is the fact that
caching the manifest is mandatory for a successful return of a manifest
struct. This patch addresses both these issues by introducing:

* a shared http client
* a result check of the http response
* and an optional cache call

Changes
=======

![IMPROVE] manifest file loading and caching