Fix adjust_path for installation on windows OS (#44)
Description
===========

The helper method `adjust_path` that checks if the proved path to a
unity installation is pointing to a Unity executable adjusted the path
incorrect. In windows the path to the executable inside of an
installation is as follows:

`installation_dir\Editor\Unity.exe`

We need to get the parent of the `Editor` directory. The incorrect
implementation did one `parent` to much.

Changes
=======

* ![FIX] `adjust_path` implementation for windows