Fix linux installation version parsing
Description
===========

The Linux editor has no official way to find the version
of an installation. Unity Hub only parses the path
to the installation and reads the version from it.
If the path contains no version string none can be read from it.
This patch attempts to read the version from
the executable with the UNIX command line tool `strings`.

Changes
=======

* ![FIX] ![LINUX] installation version parsing