Improve unity component enum
Description
===========

This is a refactoring for the `uvm_core::unity::Component` enum as a
prepertion for further patches and to increase the parity to unity hub.

This patch adds missing component values and the mapping to component
ids/installation locations.
To make the code more readable the whole module is now split into the
system specific and common part.

This patch also adds a few test cases to sanity check the `to_string`
and `from_str` implementations.

Changes
=======

![IMPROVE] `uvm_core::unity::Component`