Improve unity::version::Manifest
Description
===========

A small refactoring for `unity::version::Manifest`.
The version field `unity::Version` in this struct is now a reference.
The manifest gets created with this version and doesn't need to own it.
This patch also adds smaller additions for the contained structs in the
`manifest.rs` file and some code reorderings.

The `ComponentData` struct contains a new field `download_url` which
gets created after deserializing the manifest ini file. This allows to
fetch the download url of a component without having a reference to the
manifest itself.

As a small cleanup I created a custom serde module in `de.rs` with
functions to deserialize `bool` values in ini files.

Changes
=======

* ![IMPROVE] `unity::version::Manifest`
* ![ADD] ![NEW] field `download_url` to `ComponentData`
* ![ADD] ![NEW] module `unity::version::manifest::de` with custom serde functions