Refactor platform code into sys module
Description
===========

A few modules in `uvm_core` need to branch the implementation detail
based on the current operating system. For one-line changes this can be
easily done on code but for whole sets of functions or routines custom
modules are loaded and linked based on the os. This patch moves these
modules into a common root module called `sys`. This has the advantage
that the compile check only has to happen on the root level of the new
`sys` module. The only big downside is the duplicated folder structure.
It would be nice to move the `sys` module into a separate crates but
that would be difficult because of the cyclic dependencies.

Some platform switches still needs to be addressed in a later patch.

Changes
=======

* ![IMPROVE] move platform code into `sys` module