Add new API methods to create manifest from reader path
Description
===========

The current API of `unity::version::Manifest` allows only to load a
manifest file with a specified version from the internet. If the
manifest in question was cached the cached version was returned. This
patch adds new API's to create `unity::version::Manifest` structs from
`Read` or `Path`. Because of the internal structure of Manifest and the
problem that a manifest has no idea about it's own version we need to
parse the version before we can parse the rest of the file. This might
be worth another refactor.

Changes
=======

* ![ADD] `read_manifest_version_from_path` in `unity::version::Manifest`
* ![ADD] `read_manifest_version` in `unity::version::Manifest`
* ![ADD] `from_reader` in `unity::version::Manifest`
* ![ADD] `new` in `unity::version::Manifest`