Add modules json generation logic (#52)
Description
===========

This patch adds the main logic to generate the `modules.json` from
manifest ini files and static conventions. This solutions is brought
over from Unity Hub. These files are needed for Unity Hub so the app can
install additional components. In the future it makes sense to pull
these definitions directly from the Unity Hub release stream files.
This patch also includes a temp command to generate the modules json for
Unity installations. The files are not generated at the moment as part
of the installation process.

Changes
=======

![ADD] logic to generate `modules.json` files
![Add] temp command `uvm-generate-modules-json`