Add auxiliary command download-manifest
Description
===========

For testing purposes I created a small helper command to download a
unity manifest by version and platform. I don't know how long this
command will be here to stay and I don't see it as part of the API.

To configure the platform parameter on the `IniUrl` I created a new
builder struct. By default the current system platform will be used.
Before this was done with compile time checks. This has moved into the
default implementation of the new enum `Platform`. The `IniUrlBuilder`
can be used to set a different platform than the current system
platform.

Changes
=======

![ADD] ![NEW] auxiliary command `uvm-download-manifest`
![ADD] ![NEW] enum `Platform`
![IMPROVE] `unity::IniUrl` with new initializer `IniUrlBuilder`