Add uvm-modules command
Description
===========

As a preparation for changes in the commandline interface of the install
command, I added this new helper command `uvm-modules` which lists
available modules for the given versions.

```
uvm-modules - List available modules for a specified unity version.

Usage:
  uvm-modules [options] [--category=CATEGORY...] <version>
  uvm-modules (-h | --help)

Options:
  -c=CATEGORY, --category=CATEGORY    filter by category
  -v, --verbose                       print more output
  -d, --debug                         print debug output
  --color WHEN                        Coloring: auto, always, never [default: auto]
  -h, --help                          show this help message and exit

Arguments:
  <version>              The unity version to list modules for in the form of `2018.1.0f3`
```

The command is quite slow at the moment because no caching is involved.

To allow the category filtering I created a new enum `unity::Category`.
This enum is also used on the modules generation logic.

Changes
=======

![ADD] ![NEW] command `uvm-modules`
![ADD] enum `unity::component::Category`
![IMPROVE] module generation with new `Category` enum