Fix generate modules json command (#57)
Description
===========

The first version of the command missed the desired usage. So I split
the command into two.

- `generate-modules-json`
- `fix-modules-json`

The use case for the `fix-modules-json` command is the actual initial
idea for the `generate-modules-json` command which got lost during
implementation. This new command reads the given version to
generate/update the `modules.json` file for. In contrast to the original
implementation, it marks modules in the file as selected when they are
installed.

The `generate-modules-json` command was changed to mimic the behavior of
`download-manifest`. It is possible to generate a `modules.json` file
for any valid Unity version. The term `generate` is important as this
file is still created at runtime and not fetched from any remote or
local location.

Changes
=======

![FIX] `generate-modules-json` command
![ADD] `fix-modules-json` command