Improve module struct layout (#82)
Description
===========

To make it esier in the future to map and connect different modules
which are linked with the `parent` or `sync` field, we need to use the
same types for module `id`, `sync` and `parent`. This patch does this in
a crude way. I had some problems getting `serde` to emit the correct
code so that I can also serialize/deserialize `Option<Component>` types.

One noteworthy point is the fact that Unity serializes the `android`
component as `Android Build Support`, but only for the `sync` field. I
kept this error because I have no idea at the moment if Unity Hub works
if the value would be encoded correctly with `android`.

Changes
=======

![IMPROVE] `Module` struct layout