Improve macos module installer support (#90)
Description
===========

This is the first patch in a series to improve the module installation
support on all system and reach Unity Hub parity.
This patch concentrates on the macOS platform. It adds three new
installer type supports

* zip
* po
* dmg

Zip
---

On the macOS platform most modules are installed with `pkg` packages.
The only exceptions are the new external android build tools packages.
These are all configured as `zip` archives. The current implementation
extracts the content of the requested archives to a chooses destination
directory. The module description contains two new fields `renameFrom`
and `renameTwo`. These final operations are not part of this patch
because the install function has no context knowledge about the module
to be installed.

PO
--

`*.po` files are translation files which get installed directly into the
`Unity.app` content directory. Unity Hub can't install them via the menu
but the `modules.json` lists them.

DMG
---

*Visual Studio* is the only module at the moment which is deployed and
installed via a disk image. The implementation mounts the image and
tries to fetch a file with the postfix `.app` and copies it to the
destination location. There is no error handling or check if the
application already exists at the moment.

Missing installer types
------------------------

One known module can't be installed at the moment. The `mono` package
for *Visual Studio* is also deployed as a `pkg` installer. This
installer needs to run with the default `pkg` tool. There is no way of
discovering the differences between Unity's own modules in `pkg` format
and other external ones without giving the install function more
context.

Helper command
--------------

I added a small helper command to test these new installer types. It's a
direct copy of `uvm-install-editor` called `uvm-install-module`. It has
no realworld usage and will be removed in due time.

Changes
=======

![ADD] ![MACOS] `zip` module installer support
![ADD] ![MACOS] `dmg` module installer support
![ADD] ![MACOS] `po` language files installer support
![ADD] helper command `uvm-install-module`