Split Manifest data structs into multiple modules for sharing
I created a new module `ini` which holds a struct `IniData` which is a
copy of the old definition `ComponentData`. The new struct gets
reexported as `ComponentData` in the manifest module for backwards
compatibility.

I also moved `MD5` to a seperate module and implemented some `fmt`
traits to print the value as `HEX`.