Create new ini manifest struct
I pulled copied most of the internal manifest handling which is only
`ini` related to a seperate module. I aim to replace the ini handling
inside of `Manifest` v1 with the shared module. I still need the `ini`
logic for the `v2` implementation as a middlestep to load the version
manifest in cases of cache miss etc.