Refactor Unity Manifest (#94)
Description
===========

This is part I of an ongoing refactor to use the `modules.json`
information as base in uvm instead of the `ini` manifest information.
The reason is that the `ini` manifest is only a subset of the
`modules.json` variant. In the future I don't want to handle
the special cases of `ini` and `json` version at all.

The idea is to create a new `Manifest` struct that holds the
information provided in the `modules.json` files. Since we have no way
to load this file directly I will opt in to generate it on the fly
from the `ini` manifest. This means that the ini loading part will
only be an in-between step in the new load method.

This patch creates the ground work for adding a new Manifest struct.
All pulled out the `ini` handling logic and made it
shareable in its own module. The `Manifest` struct only delegates it
internal loading calls to this new shared struct.
This gives me the possibility to add a new `Manifest` struct
(in a new module namespace for the time being) which loads the `ini`
and generates converts it into `Modules` representation.

Changes
=======

![ADD] new module namespace `uvm_core::unity::version::manifest::v1`
![IMPROVE] reexport `v1::Manifest` to be API backwards compatible
![ADD] new `ini` module with structs to handle ini manifest loading
![IMPROVE] use shared `ini` logic from within `Manifest`

Steps
=====

* Move manifest to new namespace

This patch moves the `Manifest` struct and additional items to a new
sub namespace `v1`. The structs get reexported as before to not break
the current API.

* Split Manifest data structs into multiple modules for sharing

I created a new module `ini` which holds a struct `IniData` which is a
copy of the old definition `ComponentData`. The new struct gets
reexported as `ComponentData` in the manifest module for backwards
compatibility.

I also moved `MD5` to a seperate module and implemented some `fmt`
traits to print the value as `HEX`.

* Create new ini manifest struct

I pulled copied most of the internal manifest handling which is only
`ini` related to a seperate module. I aim to replace the ini handling
inside of `Manifest` v1 with the shared module. I still need the `ini`
logic for the `v2` implementation as a middlestep to load the version
manifest in cases of cache miss etc.

* Use new shared IniManifest as internal load logic for Manifest

This patch removes all duplicate ini handling code and uses the new
shared module from within `Manifest`.

* Run certain tests only on macOS