Improve module destination fields as std::path::PathBuf (#99)
* Fix zip module installer (#98)

Description
===========

After some additional testing with the new installer logic I realized
that some modules where not working inside of Unity. Most prominently is
the `OpenJDK` module which stopped working for me after an update to the
latest `2019.3` beta. The reason was that Unity shipped the JDK twice
and never actually used the JDK installed as a zip package but the one
contained in the base android module which is shipped as a `pkg`
package. The issue boils down to the fact that the `unzip` library isn't
setting the UNIX permissions after extracting the files from the
archive. Android SDK tools missed the executable permission and where
unusable. This patch replaces the [`unzip`] crate with [`zip`]. The whole
extract logic is converted from the example in the [`zip`] repository.

A second smaller fix was to remove the `clean_directory(destination)?;`.
Two of the new android SDK support modules share the same `destination`.
So installing one after the other ended up deleting the content from one
module all the time.

Changes
=======

* ![FIX] zip module installer

[`zip`]:	https://github.com/mvdnes/zip-rs
[`unzip`]:	https://crates.io/crates/unzip

* Improve module destination fields as std::path::PathBuf

Description
===========

This patch makes handling of file paths from modules easier. Instead of
serializing these values to `String` objects we use `std::path::PathBuf`
instead.

Changes
=======

* ![IMPROVE] module destination fields as `std::path::PathBuf`