Refactor Unity Manifest part II
Description
===========

This patch marks the second part of the manifest refactoring. The goal
is to create a new manifest struct that contains module information
loaded from the modules.json files instead of the old ini installer
files. I created a complete new `Manifest` struct in a new namespace
`v2`. I also added a reexport with the name `ManifestV2` for a more
convenient access. The bulk of the changes are refactorings to create the
modules list from both the old `Manifest` struct and the new
`InitManifest` struct.

Changes
=======

![ADD] ![NEW] `unity_core::unity::version::manifest::v2::Manifest`
![IMPROVE] module generation from `Manifest` and `IniManifest`
![ADD] reexports of new `Manifest`