Refactor Loader
Description
===========

For the new planned installer I need to slightly refactor the
`uvm_core::installer::Loader` struct and adjust it's API. The plan to
remove `install::Variant` in the long run and replace it directly with
`uvm_core::unity::Component`. The `new` method is changed to accept any
type which implements `Into<Component>`. I also realized that each
loader instance loads its own copy of the manifest. This is unnecessary.
Instead the loader is now constructed with a manifest ref instead. I
changed the `download_installer` helper method to limit the API
breakage.

Changes
=======

* ![CHANGE] `uvm_core::installer::Loader` API
* ![IMPROVE] loader manifest loading