Fix missing installer filename (#108)
Description
===========

Some of the new module packages contain installer urls without filename
parts. For these urls the tool needs to make some additional calls to
said url to retrieve the final name. This patch fix implements a fix
with the help of a new utility method.

The new function `uvm_core::utils::UrlUtils::get_file_name_from_url`
will check if the provided url ends with a filename. If yes this
filename will be returned otherwise execute a `HEAD` call and try to
read the `content-disposition` header. If this fails try to read the
name from the redirection url.

Changes
=======

* ![FIX] missing installer filename
* ![ADD] `UrlUtils` struct with `get_file_name_from_url` function