Refactor installer API (#109)
Description
===========

To support all new and old support modules I need to adjust the
installer API. At the moment each module installation required a
destination path for the installation. Sadly some modules only have no
destination configured and either use a default value or none at all.
That is a problem because validating which default value to use is
platform and installer type specific. Which means that this can only be
handled in the installer directly. I made the decision to change the API
for `install_editor` and `install_module` to support an
`Option<AsRef<Path>>` value. The installer will valid the destination
and fail if a destination is required but not passed down.

Changes
=======

* ![CHANGE] installer API