Improve windows module installer support
Description
===========

Third patch to gaiin Unity Hub parity for module installation.
For windows it was necessary to implement other changes first to get to
the point to savely implement the missing installer types.
This patch adds the missing installer types:

* zip
* po
* msi

Zip
---

Zip is used as for all other platforms. It invokes the shared cross
platform implementation.

MSI
---

The `msi` installer is tricky. It seems to be used only by `Facebook
Gameroom`. It seems Unity Hub expects a specigic `cmd` string to replace
the installer path. I use the same trick as used with the `exe`
installers by creating a temp `cmd` script which invokes the installer
comand. At one point it might be worth investigating in a `CommandExt`
to call the windows process in the correct way.

Changes
=======

![ADD] ![WINDOWS] `zip` module installer support
![ADD] ![WINDOWS] `msi` module installer support
![ADD] ![WINDOWS] `po` language files installer support
![CHANGE] ![WINDOWS] installer API and support optional `cmd` string