Add new install command uvm_install2
Description
===========

This patch brings in a complete new installer. It's a first step for an
installer that manages to be fully compatible with all unity module
types. The commandline interface is very different compared to the
original `uvm-install` command. Modules have to be references by `id`
instead of commandline flags. It becomes impossible to keep a list of
possible flags when each version of unity changes names/numbers of
modules. The helper task `uvm-modules` should be used to fetch a list of
available modules for a given unity version.

Usage
-----

```
USAGE:
    uvm-install2 [FLAGS] [OPTIONS] <version> [--] [destination]

FLAGS:
    -d, --debug
            print debug output

    -h, --help
            Prints help information

        --with-sync
            install also synced modules

    -V, --version
            Prints version information

    -v, --verbose
            print more output

OPTIONS:
        --color <color>
            Coloring [default: auto]  [possible values: auto, always, never]

    -m, --module <module>...
            A support module to install. You can list all awailable
            modules for a given version using `uvm-modules`

ARGS:
    <version>
            The unity version to install in the form of `2018.1.0f3`

    <destination>
            A directory to install the requested version to
```

The new command is written with [`clap`] instead of [`docopt`] to check
out if this way of declaring the commandline interface is more natural.

The command itself will be slower compared then the old one because it
installs the requested unity version and modules in sequentially. The
plan is to adress this after verifing that the new installer actually
works like intended.

Changes
=======

* ![ADD] ![NEW] installer `uvm-install2`

[`clap`]:  https://crates.io/crates/clap
[`docopt`]:  https://crates.io/crates/docopt