Fix modules install path (#119)
Description
===========

I realized that some of the installpath and install_location returns
were incorrect when compared to the offical `modules.json` files coming
from Unity.

Example:

_macOS android-sdk-build-tools_

```json
    "destination": "{UNITY_PATH}/PlaybackEngines/AndroidPlayer/SDK/build-tools",
    "renameTo": "{UNITY_PATH}/PlaybackEngines/AndroidPlayer/SDK/build-tools/28.0.3",
    "renameFrom": "{UNITY_PATH}/PlaybackEngines/AndroidPlayer/SDK/build-tools/android-9"
```

_windows android-sdk-build-tools_

```json
    "destination": "{UNITY_PATH}/Editor/Data/PlaybackEngines/AndroidPlayer/SDK",
    "renameTo": "{UNITY_PATH}/Editor/Data/PlaybackEngines/AndroidPlayer/SDK/platforms/android-28",
    "renameFrom": "{UNITY_PATH}/Editor/Data/PlaybackEngines/AndroidPlayer/SDK/android-9"
```

The difference is marginal and yields a similar installation path. But I
made a mistake coping the macOS values over to windows and adjusted some
values based on the expected windows output path. Classic problem. Never
trust your own fixtures.

Since I want to produce the same `modules.json` files as Unity Hub and
the modules in questions are not part of the older `ini` manifest files
I had to fix this.

I also took this chance of adjusting the `install_location` return types
slightly. In most cases the install path and `install_location` is the
same and or unique enough to just probe for the existence of the
directory to determine if something is installed or not. In case of
`android` it is not. The `destination` field points in some cases to a
shared location (`PlaybackEngines/AndroidPlayer/SDK`) even though the
final resources are located in a subdirectory. This was never a problem
because modules always had a unique install path. I took the liberty to
manually adjust this path for installation checks for now.
Means:

macOS _android-sdk-build-tools_:
* installpath `->` `{UNITY_PATH}/PlaybackEngines/AndroidPlayer/SDK`
* install_location `->` `{UNITY_PATH}/PlaybackEngines/AndroidPlayer/SDK/build-tools`

I hope this solution works until I have time to implement a check to ask
the `modules.json` if a given module is installed or not.

Changes
=======

* ![FIX] ![WINDOWS] module install path
* ![IMPROVE] module `install_location` path