Refactor uvm_install_core (#130)
Description
===========

This is a general overhaul for the installer logic. Instead of
validating the installer model during execution, this new logic splits
the creation and execution into two steps. This patch adds a generic
`Installer` struct which contains in it's most generic form some
boilerplate creation and bootstrapping function. Each platform and
installer type implements custom installation logic. A helper function
validates the raw input from the manifest and instantiates an installer
object. If the provided parameters are invalid the installation object
can't be created.

Changes
=======

![IMPROVE] `uvm_install_core` crate