Use structops instead of docopts for uvm-launch
Description
===========

This is a test patch to see how difficult it is to use
structops instead of docopts for the cli commands. I decided to
go with structops over clap since clap 3 is basically structops.
Moving to clap version 3 which is still in beta will be easier this
way.

I moved some generic code that I will need for other commands
into `uvm_cli` so I only need this as a dependency.

Changes
=======

* ![IMPROVE] use `structops` over `docopts` in `uvm-launch` command