Fix missing install process lock for install2 command
Description
===========

The `install2` command was not locking the install
process for the given version like the old implementation does.
That results in some buggy behavior when two processes of uvm install2
try to install or alter the same unity installation.

The patch simply locks the whole install method if another
process is trying to alter the Unity installation for a given
version. This means even parallel calls of `install` and `install2`
should be covered.

Changes
=======

* ![FIX] missing install process lock for `install2` command